Fix file rejection issues
If you have for example a minimum size defined and the added image causes an error your code will save the image even thought the original method will reject it afterwards. This is resulting in 0x0 images being safed to the database while the file will be unlink'ed. (ProcessWire will tell that it is unable to display the image)